02_workflows_and_agents/studio/requirements.txt
langgraph>=0.2.0
langsmith>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
ipykernel
python-dotenv
langchain-tavily
langgraph-cli[inmem]
langchain-anthropic
langchain-community
beautifulsoup4
